+++ asf_search/constants/DATASET.py
@@ -17,5 +17,13 @@ SEASAT = 'SEASAT'
 NISAR = 'NISAR'
 """NISAR provides L and S-band SAR data to measure Earth's changing ecosystems,
  dynamic surfaces, and ice masses with 12-day regularity
- on ascending and descending passes."""
+ on ascending and descending passes.
+ See https://nisar-docs.asf.alaska.edu/products-overview/ 
+ for details on L0B-L3 science and supporting data product types
+ """
+NISAR_SCIENCE_PRODUCTS = 'NISAR-SCIENCE-PRODUCTS'
+"""Convenient subset of the main `NISAR` dataset with just L0B, L1, L2, and L3 science products. 
+See https://nisar-docs.asf.alaska.edu/products-overview/ for more information on NISAR science products.
+(For all other NISAR product types including supporting products, such as orbit ephemeras and water masks)
+use the regular `NISAR` dataset."""
 TROPO = 'TROPO'